requirements.txt
fastapi
uvicorn
websockets
python-multipart
cantools
pandas
pyserial
